refactor: big cleanup, folder and file renaming
