+++ .gitignore
@@ -1,7 +1,6 @@
 # Include your project-specific ignores in this file
 # Read about how to use .gitignore: https://help.github.com/articles/ignoring-files
-dist
-bower_components
+public
 node_modules
 npm-debug.log
 vendor
