--- .nvmrc
@@ -1,1 +1,0 @@
-8.11.1
