--- .travis.yml
@@ -1,35 +1,0 @@
-sudo: false
-language: php
-php:
-  - nightly
-  - 5.6
-  - 5.5
-  - 5.4
-
-env:
-  - TRAVIS_NODE_VERSION="4"
-
-matrix:
-  allow_failures:
-    - php: nightly
-
-cache:
-  directories:
-    - bower_components
-    - node_modules
-
-install:
-  - nvm install $TRAVIS_NODE_VERSION
-  - nvm use $TRAVIS_NODE_VERSION
-  - npm install -g npm@latest
-  - npm install -g bower gulp jscs
-  - npm install
-  - composer self-update && composer --version
-  - export PATH="$HOME/.composer/vendor/bin:$PATH"
-  - composer global require squizlabs/php_codesniffer
-
-script:
-  - npm run build
-  - npm run jshint
-  - npm run jscs
-  - phpcs --standard=ruleset.xml --extensions=php -n -s .
