+++ app/classes/ThemeActions.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Vibrato\Classes;
+
+class ThemeActions
+{
+    public function init()
+    {
+    }
+}
