+++ app/classes/WalkerNav.php
@@ -1,11 +1,12 @@
 <?php
-namespace ZGM\Vibrato\Classes;
+
+namespace Vibrato\Classes;
 
 use Walker_Nav_Menu;
 
 class WalkerNav extends Walker_Nav_Menu
 {
-        /**
+    /**
      * What the class handles.
      *
      * @since 3.0.0
@@ -13,8 +14,8 @@ class WalkerNav extends Walker_Nav_Menu
      *
      * @see Walker::$tree_type
      */
-    public $tree_type = array( 'post_type', 'taxonomy', 'custom' );
- 
+    public $tree_type = array('post_type', 'taxonomy', 'custom');
+
     /**
      * Database fields to use.
      *
@@ -28,7 +29,7 @@ class WalkerNav extends Walker_Nav_Menu
         'parent' => 'menu_item_parent',
         'id'     => 'db_id',
     );
- 
+
     /**
      * Starts the list before the elements are added.
      *
@@ -40,19 +41,20 @@ class WalkerNav extends Walker_Nav_Menu
      * @param int      $depth  Depth of menu item. Used for padding.
      * @param stdClass $args   An object of wp_nav_menu() arguments.
      */
-    public function start_lvl( &$output, $depth = 0, $args = array() ) {
-        if ( isset( $args->item_spacing ) && 'discard' === $args->item_spacing ) {
+    public function start_lvl(&$output, $depth = 0, $args = array())
+    {
+        if (isset($args->item_spacing) && 'discard' === $args->item_spacing) {
             $t = '';
             $n = '';
         } else {
             $t = "\t";
             $n = "\n";
         }
-        $indent = str_repeat( $t, $depth );
- 
+        $indent = str_repeat($t, $depth);
+
         // Default class.
-        $classes = array( 'sub-menu' );
- 
+        $classes = array('sub-menu');
+
         /**
          * Filters the CSS class(es) applied to a menu list element.
          *
@@ -62,12 +64,12 @@ class WalkerNav extends Walker_Nav_Menu
          * @param stdClass $args    An object of `wp_nav_menu()` arguments.
          * @param int      $depth   Depth of menu item. Used for padding.
          */
-        $class_names = join( ' ', apply_filters( 'nav_menu_submenu_css_class', $classes, $args, $depth ) );
-        $class_names = $class_names ? ' class="' . esc_attr( $class_names ) . '"' : '';
- 
+        $class_names = join(' ', apply_filters('nav_menu_submenu_css_class', $classes, $args, $depth));
+        $class_names = $class_names ? ' class="' . esc_attr($class_names) . '"' : '';
+
         $output .= "{$n}{$indent}<ul$class_names>{$n}";
     }
- 
+
     /**
      * Ends the list of after the elements are added.
      *
@@ -79,18 +81,19 @@ class WalkerNav extends Walker_Nav_Menu
      * @param int      $depth  Depth of menu item. Used for padding.
      * @param stdClass $args   An object of wp_nav_menu() arguments.
      */
-    public function end_lvl( &$output, $depth = 0, $args = array() ) {
-        if ( isset( $args->item_spacing ) && 'discard' === $args->item_spacing ) {
+    public function end_lvl(&$output, $depth = 0, $args = array())
+    {
+        if (isset($args->item_spacing) && 'discard' === $args->item_spacing) {
             $t = '';
             $n = '';
         } else {
             $t = "\t";
             $n = "\n";
         }
-        $indent  = str_repeat( $t, $depth );
+        $indent  = str_repeat($t, $depth);
         $output .= "$indent</ul>{$n}";
     }
- 
+
     /**
      * Starts the element output.
      *
@@ -105,19 +108,20 @@ class WalkerNav extends Walker_Nav_Menu
      * @param stdClass $args   An object of wp_nav_menu() arguments.
      * @param int      $id     Current item ID.
      */
-    public function start_el( &$output, $item, $depth = 0, $args = array(), $id = 0 ) {
-        if ( isset( $args->item_spacing ) && 'discard' === $args->item_spacing ) {
+    public function start_el(&$output, $item, $depth = 0, $args = array(), $id = 0)
+    {
+        if (isset($args->item_spacing) && 'discard' === $args->item_spacing) {
             $t = '';
             $n = '';
         } else {
             $t = "\t";
             $n = "\n";
         }
-        $indent = ( $depth ) ? str_repeat( $t, $depth ) : '';
- 
-        $classes   = empty( $item->classes ) ? array() : (array) $item->classes;
+        $indent = ($depth) ? str_repeat($t, $depth) : '';
+
+        $classes   = empty($item->classes) ? array() : (array) $item->classes;
         $classes[] = 'menu-item-' . $item->ID;
- 
+
         /**
          * Filters the arguments for a single nav menu item.
          *
@@ -127,8 +131,8 @@ class WalkerNav extends Walker_Nav_Menu
          * @param WP_Post  $item  Menu item data object.
          * @param int      $depth Depth of menu item. Used for padding.
          */
-        $args = apply_filters( 'nav_menu_item_args', $args, $item, $depth );
- 
+        $args = apply_filters('nav_menu_item_args', $args, $item, $depth);
+
         /**
          * Filters the CSS classes applied to a menu item's list item element.
          *
@@ -140,9 +144,9 @@ class WalkerNav extends Walker_Nav_Menu
          * @param stdClass $args    An object of wp_nav_menu() arguments.
          * @param int      $depth   Depth of menu item. Used for padding.
          */
-        $class_names = join( ' ', apply_filters( 'nav_menu_css_class', array_filter( $classes ), $item, $args, $depth ) );
-        $class_names = $class_names ? ' class="' . esc_attr( $class_names ) . '"' : '';
- 
+        $class_names = join(' ', apply_filters('nav_menu_css_class', array_filter($classes), $item, $args, $depth));
+        $class_names = $class_names ? ' class="' . esc_attr($class_names) . '"' : '';
+
         /**
          * Filters the ID applied to a menu item's list item element.
          *
@@ -154,22 +158,22 @@ class WalkerNav extends Walker_Nav_Menu
          * @param stdClass $args    An object of wp_nav_menu() arguments.
          * @param int      $depth   Depth of menu item. Used for padding.
          */
-        $id = apply_filters( 'nav_menu_item_id', 'menu-item-' . $item->ID, $item, $args, $depth );
-        $id = $id ? ' id="' . esc_attr( $id ) . '"' : '';
- 
+        $id = apply_filters('nav_menu_item_id', 'menu-item-' . $item->ID, $item, $args, $depth);
+        $id = $id ? ' id="' . esc_attr($id) . '"' : '';
+
         $output .= $indent . '<li' . $id . $class_names . '>';
- 
+
         $atts           = array();
-        $atts['title']  = ! empty( $item->attr_title ) ? $item->attr_title : '';
-        $atts['target'] = ! empty( $item->target ) ? $item->target : '';
-        if ( '_blank' === $item->target && empty( $item->xfn ) ) {
+        $atts['title']  = !empty($item->attr_title) ? $item->attr_title : '';
+        $atts['target'] = !empty($item->target) ? $item->target : '';
+        if ('_blank' === $item->target && empty($item->xfn)) {
             $atts['rel'] = 'noopener noreferrer';
         } else {
             $atts['rel'] = $item->xfn;
         }
-        $atts['href']         = ! empty( $item->url ) ? $item->url : '';
+        $atts['href']         = !empty($item->url) ? $item->url : '';
         $atts['aria-current'] = $item->current ? 'page' : '';
- 
+
         /**
          * Filters the HTML attributes applied to a menu item's anchor element.
          *
@@ -189,19 +193,19 @@ class WalkerNav extends Walker_Nav_Menu
          * @param stdClass $args  An object of wp_nav_menu() arguments.
          * @param int      $depth Depth of menu item. Used for padding.
          */
-        $atts = apply_filters( 'nav_menu_link_attributes', $atts, $item, $args, $depth );
- 
+        $atts = apply_filters('nav_menu_link_attributes', $atts, $item, $args, $depth);
+
         $attributes = '';
-        foreach ( $atts as $attr => $value ) {
-            if ( ! empty( $value ) ) {
-                $value       = ( 'href' === $attr ) ? esc_url( $value ) : esc_attr( $value );
+        foreach ($atts as $attr => $value) {
+            if (!empty($value)) {
+                $value       = ('href' === $attr) ? esc_url($value) : esc_attr($value);
                 $attributes .= ' ' . $attr . '="' . $value . '"';
             }
         }
- 
+
         /** This filter is documented in wp-includes/post-template.php */
-        $title = apply_filters( 'the_title', $item->title, $item->ID );
- 
+        $title = apply_filters('the_title', $item->title, $item->ID);
+
         /**
          * Filters a menu item's title.
          *
@@ -212,14 +216,14 @@ class WalkerNav extends Walker_Nav_Menu
          * @param stdClass $args  An object of wp_nav_menu() arguments.
          * @param int      $depth Depth of menu item. Used for padding.
          */
-        $title = apply_filters( 'nav_menu_item_title', $title, $item, $args, $depth );
- 
+        $title = apply_filters('nav_menu_item_title', $title, $item, $args, $depth);
+
         $item_output  = $args->before;
         $item_output .= '<a' . $attributes . '>';
         $item_output .= $args->link_before . $title . $args->link_after;
         $item_output .= '</a>';
         $item_output .= $args->after;
- 
+
         /**
          * Filters a menu item's starting output.
          *
@@ -234,9 +238,9 @@ class WalkerNav extends Walker_Nav_Menu
          * @param int      $depth       Depth of menu item. Used for padding.
          * @param stdClass $args        An object of wp_nav_menu() arguments.
          */
-        $output .= apply_filters( 'walker_nav_menu_start_el', $item_output, $item, $depth, $args );
+        $output .= apply_filters('walker_nav_menu_start_el', $item_output, $item, $depth, $args);
     }
- 
+
     /**
      * Ends the element output, if needed.
      *
@@ -249,8 +253,9 @@ class WalkerNav extends Walker_Nav_Menu
      * @param int      $depth  Depth of page. Not Used.
      * @param stdClass $args   An object of wp_nav_menu() arguments.
      */
-    public function end_el( &$output, $item, $depth = 0, $args = array() ) {
-        if ( isset( $args->item_spacing ) && 'discard' === $args->item_spacing ) {
+    public function end_el(&$output, $item, $depth = 0, $args = array())
+    {
+        if (isset($args->item_spacing) && 'discard' === $args->item_spacing) {
             $t = '';
             $n = '';
         } else {
