--- assets/scripts/customizer.js
@@ -1,8 +1,0 @@
-(function($) {
-  // Site title
-  wp.customize('blogname', function(value) {
-    value.bind(function(to) {
-      $('.brand').text(to);
-    });
-  });
-})(jQuery);
