--- assets/styles/common/_global.scss
@@ -1,3 +1,0 @@
-a,button,input[type="submit"]{
-	transition: all 0.2s;
-}
