--- assets/styles/common/_ie.scss
@@ -1,3 +1,0 @@
-@include ie11(){
-	//
-}
