--- assets/styles/common/_mixins.scss
@@ -1,5 +1,0 @@
-@mixin ie11(){
-  @media screen and (-ms-high-contrast: active), screen and (-ms-high-contrast: none) {  
-    @content; 
-  }
-}
