--- assets/styles/common/_variables.scss
@@ -1,12 +1,0 @@
-//Font Awesome
-$font-primary: 'Open Sans', sans-serif;
-$font-icons: "Font Awesome 5 Free"; // note: this wants font-weight: 900 because FontAwesome is trying to be sneaky
-$font-brands: "Font Awesome 5 Brands";
-$icon-plus: '\f055';
-$icon-minus: '\f056';
-$icon-minus-solid: '\f068';
-$icon-plus-solid: '\f067';
-$icon-arrow-left: '\f060';
-$icon-arrow-right: '\f061';
-$icon-arrow-left-circle: '\f137';
-$icon-arrow-right-circle: '\f138';
