--- assets/styles/components/_accordion.scss
@@ -1,53 +1,0 @@
-.accordion{
-	position: relative;
-	&-heading{
-		@apply py-8 px-8 flex items-center justify-between w-full uppercase border-gray-300 border-t-4;
-    //Accordion plus / minus and animation
-			&.is-active{
-				background-color: #fff;
-				&:before{
-					opacity: 1;
-					transform: rotate(0);
-				}
-				&:after{
-					opacity: 0;
-					transform: rotate(0);
-				}
-			}
-			&:before,
-			&:after{
-				font-family: $font-icons;
-				@apply font-bold;
-				font-style: normal;
-	    	font-variant: normal;
-	    	text-rendering: auto;
-	    	-webkit-font-smoothing: antialiased;
-	    	transition: opacity 0.3s;
-	    	transition: transform 0.2s;
-	    	position: absolute;
-	    	left: 1rem;
-			}
-			&:before{
-				content: $icon-minus-solid;
-				opacity: 0;
-				transform: rotate(-90deg);
-			}
-			&:after{
-				content: $icon-plus-solid;
-				transform: rotate(-90deg);
-			}
-	}
-	&-content{
-		z-index: -1;
-		opacity: 0;
-		background-color: white;
-		max-height: 0;
-		overflow: hidden;
-		transition: max-height 0.3s ease-out;
-		&.is-active{
-			z-index: 1;	
-			padding: 15px 1rem;
-			opacity: 1;
-		}
-	}
-}
