--- assets/styles/main.scss
@@ -1,26 +1,0 @@
-//Tailwindcss https://tailwindcss.com/
-//Tailwindcss - Base
-@import "~tailwindcss/base";
-//Tailwindcss - Components
-@import "~tailwindcss/components";
-
-
-//Local scss files run before everything
-@import "common/variables";
-@import "overrides/hamburgers";
-
-//Node Modules
-@import "~hamburgers/_sass/hamburgers/hamburgers";
-
-//Local scss files
-@import "common/mixins";
-@import "common/global";
-@import "common/utilities";
-@import "common/ie";
-@import "common/type";
-@import "components/buttons";
-@import "components/links";
-@import "components/accordion";
-
-//Tailwindcss Utilities
-@import "~tailwindcss/utilities";
