--- assets/styles/overrides/_hamburgers.scss
@@ -1,8 +1,0 @@
-/** Hamburger icon */
-$hamburger-padding-x: 0;
-$hamburger-padding-y: 0;
-$hamburger-layer-width: 25px;
-$hamburger-layer-height: 2px;
-$hamburger-layer-spacing: 4px;
-$hamburger-layer-color: #333;
-$hamburger-types: (squeeze);
