+++ composer.json
@@ -20,17 +20,17 @@
   "require": {
     "php": ">=7.1.0",
     "composer/installers": "~1.0",
-    "htmlburger/carbon-fields": "^3.0.1"
+    "htmlburger/carbon-fields": "^3.2"
   },
   "config": {
     "optimize-autoloader": true
   },
   "autoload": {
     "psr-4": {
-        "ZGM\\Vibrato\\": "lib/"
+        "Vibrato\\": "app/"
     },
     "classmap": [
-        "lib/"
+        "app/"
     ]
   }
 }
