+++ package.json
@@ -1,48 +1,20 @@
 {
-  "name": "sage-vibrato",
-  "version": "1.0.0",
-  "author": "Jon Waldstein <jpwaldstein@gmail.com>",
-  "homepage": "https://jonwaldstein.com",
   "private": true,
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/Jpwaldstein/vibrato.git"
-  },
-  "bugs": {
-    "url": "https://github.com/Jpwaldstein/vibrato/issues"
-  },
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "http://opensource.org/licenses/MIT"
-    }
-  ],
   "scripts": {
-    "dev": "NODE_ENV=development webpack --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "watch": "NODE_ENV=development webpack --watch --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "hot": "NODE_ENV=development webpack-dev-server --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "production": "NODE_ENV=production webpack --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
-  },
-  "engines": {
-    "node": ">= 8.11.1"
+    "dev": "npm run development",
+    "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "watch": "npm run development -- --watch",
+    "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "prod": "npm run production",
+    "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
   },
   "devDependencies": {
-    "axios": "^0.19.0",
-    "hamburgers": "^1.1.3",
-    "laravel-mix": "^4.1.2",
-    "resolve-url-loader": "2.3.1",
-    "sass": "^1.22.6",
-    "sass-loader": "7.*"
+    "autoprefixer": "^10.2.4",
+    "postcss": "^8.2.4",
+    "tailwindcss": "^2.0.2"
   },
   "dependencies": {
-    "@wordpress/dom-ready": "^2.4.0",
-    "es6-promise": "^4.2.8",
-    "jquery": "^3.4.1",
-    "object-fit-images": "^3.2.4",
-    "smooth-scroll": "^16.1.0",
-    "tailwindcss": "^1.0.5",
-    "vue": "^2.5.16",
-    "vue-template-compiler": "^2.5.16",
-    "vuex": "^3.0.1"
+    "cross-env": "^7.0.3",
+    "laravel-mix": "^6.0.11"
   }
 }
