+++ page.php
@@ -1,8 +1,9 @@
 <?php
-use ZGM\Vibrato\Classes\Theme;
+
+use Vibrato\Classes\Theme;
 ?>
 <?php while (have_posts()) : the_post(); ?>
 
   <?php get_template_part('templates/global/content', 'page'); ?>
-  
+
 <?php endwhile; ?>
