+++ tailwind.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+purge: [
+    './**/*.php',
+    './**/*.js',
+],
+  darkMode: false, // or 'media' or 'class'
+  theme: {
+    extend: {},
+  },
+  variants: {
+    extend: {},
+  },
+  plugins: [],
+}
