+++ webpack.mix.js
@@ -1,29 +1,17 @@
-const tailwindcss = require('tailwindcss');
 let mix = require('laravel-mix');
-mix.setPublicPath('./');
-mix.setResourceRoot('../');
+let path = require('path');
 
-/*
- |--------------------------------------------------------------------------
- | Mix Asset Management
- |--------------------------------------------------------------------------
- |
- | Mix provides a clean, fluent API for defining some Webpack build steps
- | for your Laravel application. By default, we are compiling the Sass
- | file for your application, as well as bundling up your JS files.
- |
- */
+mix.alias({
+    '@': path.join(__dirname, 'resources/js')
+});
 
-mix.js('assets/scripts/main.js', 'dist/scripts/')
-   .sass('assets/styles/main.scss', 'dist/styles/')
-   .copy('assets/fonts/', 'dist/fonts/')
-   .copy('assets/images/', 'dist/images/')
-   .options({
-    processCssUrls: false,
-    postCss: [
-      tailwindcss('assets/scripts/tailwind.config.js')
-    ],
-  });
+mix.js("resources/js/app.js", "public/js")
+    .copyDirectory('resources/fonts/', 'public/fonts/')
+    .copyDirectory('resources/images/', 'public/images/')
+    .postCss("resources/css/app.css", "public/css", [
+        require("tailwindcss"),
+    ])
+    .setPublicPath('public');
 
 if (!mix.inProduction()) {
     mix.webpackConfig({
